% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anderson_anova}
\alias{anderson_anova}
\title{Anderson's PERMANOVA for objects in a general metric space}
\usage{
anderson_anova(distmat, groupsizes, nperm = 999, alpha = 0.05)
}
\arguments{
\item{distmat}{the full distance matrix between all data objects, arranged according to
group membership (first all objects of group 1, then all objects of group 2, etc.)}

\item{groupsizes}{the vector of group sizes (length \eqn{k}, sums to \eqn{n}).}

\item{nperm}{the number of permutations to use for the test. The p-value is honest but random.
A higher \code{nperm} reduces its variance.}

\item{alpha}{the (significance) level of the test.}
}
\value{
A list with components
\itemize{
\item \code{Fstat} the value of the test statistic
\item \code{pval} the Monte Carlo p-value based on random permutations
\item \code{reject} logical. The rejection decision.
}
}
\description{
Perform a permutation test to decide whether \eqn{k \geq 2} groups of objects (\eqn{n} in total)
all come from the same distribution. The test is mainly sensitive to differing
locations of elements across the groups.
}
\details{
See Müller, Schuhmacher and Mateu (2023+). ANOVA for Data in Metric Spaces, with Applications to Spatial Point Patterns.
\emph{ArXiv preprint} \doi{10.48550/arXiv.2201.08664}.
}
\examples{
k <- 2
n <- 10
x <- rnorm(n)
y <- rnorm(n, 1)
distmat  <- as.matrix(dist(c(x,y)))
groupsizes <- rep(n, k)
anderson_anova(distmat, groupsizes, nperm=9999) 
# because we have univariate data from shifted normal distributions
# this is up to the random permutations the same as a t-test:
res <- t.test(x, y, var.equal=TRUE) 
res
res$statistic^2
}
\seealso{
\code{\link{msm_levene}} for a test that is sensitive to differing scatter across groups.
}
