% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{msm_levene}
\alias{msm_levene}
\title{The Levene's test from Müller, Schuhmacher and Mateu (2022) for objects in a general metric space}
\usage{
msm_levene(
  distmat,
  n1,
  n2,
  type = c("balanced", "unbalanced"),
  nperm = 999,
  alpha = 0.05
)
}
\arguments{
\item{distmat}{the full distance matrix between all data objects, arranged according to
group membership (first all objects of group 1, then all objects of group 2, etc.).}

\item{n1, n2}{group sizes (\eqn{n1+n2=n}).}

\item{type}{either "balanced" or "unbalanced". Determines which statistic to use for the test.
If \code{n1} = \code{n2} the statistics are equivalent. If n1\verb{, }n2` differ a lot
the unbalanced type is recommended.}

\item{nperm}{the number of permutations to use for the test. The p-value is honest but random.
A higher \code{nperm} reduces its variance.}

\item{alpha}{the (significance) level of the test.}
}
\value{
A list with components
\itemize{
\item \code{Fstat} the value of the test statistic
\item \code{pval} the Monte Carlo p-value based on random permutations
\item \code{reject} logical. The rejection decision.
}
}
\description{
Perform a permutation test to decide whether two groups of objects (\eqn{n} in total)
both come from the same distribution. The test is sensitive to differing scatter of elements
across the groups, but not at all to differing locations.
}
\details{
See Müller, Schuhmacher and Mateu (2023+). ANOVA for Data in Metric Spaces, with Applications to Spatial Point Patterns.
\emph{ArXiv preprint} \doi{10.48550/arXiv.2201.08664}.
}
\examples{
k <- 2
n <- 10
x <- rnorm(n)
y <- rnorm(n, 0, 3)
distmat  <- as.matrix(dist(c(x,y)))
groupsizes <- rep(n, k)
msm_levene(distmat/2, n, n, "balanced", nperm=9999, alpha=0.05) 
msm_levene(distmat/2, n, n, "unbalanced", nperm=9999, alpha=0.05)
# results are the same because groups have the same size
}
\seealso{
\code{\link{anderson_anova}} for a test that is sensitive to differing locations across groups.
}
