% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rspatER}
\alias{rspatER}
\title{Sample a spatial Erdős--Rényi graph with underlying Poisson or Binomial point process}
\usage{
rspatER(lambda, n = NULL, p = 0.5)
}
\arguments{
\item{lambda}{a non-negative number. The intensity of the underlying Poisson process.}

\item{n}{a non-negative integer. The number of points in the underlying Binomial process.
If \code{n} is specified \code{lambda} is ignored.}

\item{p}{a number \eqn{\in [0,1]}. The edge probability.}
}
\value{
the sampled spatial ER graph, as an object of class \code{"igraph"}.
}
\description{
Sample a spatial Erdős--Rényi graph with underlying Poisson or Binomial point process
}
\examples{
g <- rspatER(n=8,p=.3)
igraph::plot.igraph(g, rescale=FALSE, asp=1, xlim=c(0,1), ylim=c(0,1), vertex.size=8)
box()
}
