% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{spatgraph}
\alias{spatgraph}
\title{Create a Spatial Graph (igraph with coordinates)}
\usage{
spatgraph(x, y, from, to, sigma_adj = NULL)
}
\arguments{
\item{x}{vector of x-coordinates of the vertices.}

\item{y}{vector of y-coordinates of the vertices.}

\item{from}{vector of indices from which to draw edges.}

\item{to}{vector of indices to which to draw edges.}

\item{sigma_adj}{the adjaceny matrix}
}
\value{
an object of class \code{"igraph"}.
}
\description{
Create a Spatial Graph (igraph with coordinates)
}
\examples{
g1 <- spatgraph(c(-1,1,0), c(0,0,sqrt(3)), from=c(1,1), to=c(2,3))
g2 <- spatgraph(c(-0.5,0.5,0), c(0.5/sqrt(3),0.5/sqrt(3),2/sqrt(3)), from=c(1,2), to=c(2,3))
plot(g1, rescale=FALSE, asp=1, xlim=c(-1,1), ylim=c(0,sqrt(3)))  
plot(g2, add=TRUE, rescale=FALSE, vertex.color="lightblue")
}
